Update cleaner.py
The stopwords portion is functional. Some stopwords were not being omitted because of punctuation and the work around I found for that seems to have removed punctuation altogether. I'm not sure if that's okay..

Attempted to stem, but random words are being cut off. November is being changed to Novemb, not sure how to fix it.